feat: collection, ithem, tag Controllers almost complete

+++ CollectionManager/Controllers/UserCollectionController.cs
@@ -1,18 +1,20 @@
 ﻿using CollectionManager.Models.Domain;
 using CollectionManager.Repositories.Abstract;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace CollectionManager.Controllers
 {
-    public class UserCollectionsController : Controller
+    [Authorize]
+    public class UserCollectionController : Controller
     {
         private readonly IAccessService _accessService;
         private readonly ICollectionService _collectionService;
         private readonly ITopicService _topicService;
         private readonly UserManager<User> _userManager;
-        public UserCollectionsController
+        public UserCollectionController
             (
             ICollectionService collectionService, 
             ITopicService topicService, 
@@ -26,58 +28,72 @@ namespace CollectionManager.Controllers
             _userManager = userManager;
         }
 
-        public IActionResult Index(string userName)
+        public IActionResult Index(string authorName)
         {
-            ViewBag.UserName = userName;
-            return View(_collectionService.GetUserCollections(userName));
+            ViewBag.AuthorName = authorName;
+            return View(_collectionService.GetUserCollections(authorName));
         }
-        public IActionResult Add(string userName)
+        public IActionResult Add(string authorName)
         {
-            SetDataForAddCollection(userName);
+            SetDataForAddCollection(authorName);
             return View();
         }
         [HttpPost]
         public IActionResult Add(Collection model)
         {
-            //if (!IsUserAcess())
-            //    return Redirect("/Identity/Account/AccessDenied");
+            string authorName = _userManager.FindByIdAsync(model.UserId).Result.UserName;
+            if (!IsUserAccess(authorName))
+                return Redirect("/Identity/Account/AccessDenied");
+            SetDataForAddCollection(authorName);
+            if (!ModelState.IsValid)
+            {
+                return View(model);
+            }
             var result = _collectionService.Add(model);
-            string userName = _userManager.FindByIdAsync(model.UserId).Result.UserName;
-            SetDataForAddCollection(userName);
             if (result)
             {
                 TempData["msg"] = "Added Successfully";
-                return Redirect("/UserCollections/Add?userName="+userName);
+                return Redirect($"/UserCollection/Add?authorName={authorName}");
             }
             TempData["msg"] = "Error has occured on server side";
-            return Redirect("/UserCollections/Add?userName=" + userName);
+            return Redirect($"/UserCollection/Add?authorName={authorName}");
         }
-        public IActionResult Delete(string id, string userName)
+        public IActionResult Delete(string id, string authorName)
         {
-            //if (!IsAdminAcess())
-            //    return Redirect("/Identity/Account/AccessDenied");
+            if (!IsUserAccess(authorName))
+                return Redirect("/Identity/Account/AccessDenied");
             var result = _collectionService.Delete(id);
             if (!result) 
                 TempData["msg"] = "Deletion error";
-            return Redirect($"/UserCollections/Index?userName={userName}");
+            return Redirect($"/UserCollection/Index?authorName={authorName}");
         }
-        private void SetDataForAddCollection(string userName)
+        private void SetDataForAddCollection(string authorName)
         {
-            ViewBag.UserId = _userManager.FindByEmailAsync(userName).Result.Id;
-            ViewBag.UserName = userName;
+            ViewBag.AuthorId = _userManager.FindByEmailAsync(authorName).Result.Id;
+            ViewBag.authorName = authorName;
             SelectList topicSeectList = _topicService.GetSelectList();
             ViewBag.Topics = topicSeectList;
             ViewBag.NamesGroupOptionalFields = _collectionService.GetNamesGroupOptionalFields();
             ViewBag.GetCountOptionalFieldsInGroup = _collectionService.GetCountOptionalFieldsInGroup();
         }
 
-        private bool IsUserAcess()
+        private bool IsUserAccess(string authorName)
         {
             try
             {
-                if (!_accessService.IsUserRule(User.Identity.Name).Result)
+                if (authorName == User.Identity?.Name)
                 {
-                    return false;
+                    if (!_accessService.IsUserRule(User.Identity?.Name).Result)
+                    {
+                        return false;
+                    }
+                }
+                else
+                {
+                    if (!_accessService.IsAdminRule(User.Identity?.Name).Result)
+                    {
+                        return false;
+                    }
                 }
                 return true;
             }
